+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.AppCompat.DayNight.DarkActionBar">
+        android:theme="@style/Theme.Design.NoActionBar">
         <activity
             android:name=".LoginActivity"
             android:label="@string/title_activity_login"
