+++ app/src/main/res/drawable/round_shape.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <corners
+        android:radius="30dp"></corners>
+    <solid
+        android:color="#C0DD6F"
+        ></solid>
+</shape>
