+++ app/src/main/res/layout/fragment_home.xml
@@ -3,27 +3,39 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     tools:context=".Fragments.HomeFragment"
+    android:background="#008EB3"
     android:id="@+id/xxx"
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
-    <TextView
-        android:layout_gravity="center"
-        android:id="@+id/textView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/title_home"
-        />
-<ScrollView android:layout_height="wrap_content"
-    android:layout_width="match_parent">
-    <LinearLayout
-        android:id="@+id/filed_container"
-        android:layout_width="match_parent"
-        android:orientation="vertical"
-        android:layout_height="wrap_content"
-        >
 
-    </LinearLayout>
-</ScrollView>
+    <EditText
+        android:layout_width="match_parent"
+        android:layout_height="41dp"
+        android:layout_margin="20dp"
+        android:background="@drawable/round_button"
+        android:hint="Search...." />
+   <LinearLayout android:layout_width="match_parent" android:layout_height="wrap_content">
+
+       <ImageButton
+           android:layout_width="203dp"
+           android:layout_height="203dp"
+           android:layout_margin="10dp"
+           android:layout_weight="1"
+           android:adjustViewBounds="true"
+           android:background="@drawable/round_shape"
+           android:scaleType="fitCenter"
+           android:src="@drawable/milk" />
+       <ImageButton android:layout_width="200dp"
+           android:layout_height="200dp"
+           android:layout_weight="1"
+           android:layout_margin="10dp"
+           android:background="@drawable/round_shape"
+           android:scaleType="fitCenter"
+           android:adjustViewBounds="true"
+           android:src="@drawable/fruit"/>
+   </LinearLayout>
+
+
 
 </LinearLayout>
